blog_migration_package/automation/requirements.txt
yfinance
duckduckgo-search
google-generativeai
beautifulsoup4
curl_cffi
pandas
python-dotenv
